% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idconv.R
\name{IDX_to_IDY}
\alias{ENSEMBL_to_SYMBOL}
\alias{ENTREZID_to_SYMBOL}
\alias{IDX_to_IDY}
\alias{REFSEQ_to_SYMBOL}
\alias{SYMBOL_to_ENSEMBL}
\alias{SYMBOL_to_ENTREZID}
\alias{SYMBOL_to_REFSEQ}
\title{List available fastqc analysis modules and their reported status}
\usage{
IDX_to_IDY(IDX, IDY, ids)

SYMBOL_to_ENTREZID(ids)

ENTREZID_to_SYMBOL(ids)

SYMBOL_to_ENSEMBL(ids)

ENSEMBL_to_SYMBOL(ids)

SYMBOL_to_REFSEQ(ids)

REFSEQ_to_SYMBOL(ids)
}
\arguments{
\item{IDX}{input id type (e.g. "SYMBOL")}

\item{IDY}{output id type (e.g. "ENSEMBL")}

\item{ids}{vector of input identifiers to be converted to output identifiers}
}
\value{
A vector of converted identifiers with same length as input identifiers
}
\description{
\code{IDX_to_IDY} converts ids from identifier X to identifier Y using AnnotationDbi
}
\section{Functions}{
\itemize{
\item \code{SYMBOL_to_ENTREZID}: convert from SYMBOL to ENTREZID

\item \code{ENTREZID_to_SYMBOL}: convert from ENTREZID to SYMBOL

\item \code{SYMBOL_to_ENSEMBL}: convert from SYMBOL to ENSEMBL

\item \code{ENSEMBL_to_SYMBOL}: convert from ENSEMBL to SYMBOL

\item \code{SYMBOL_to_REFSEQ}: convert from SYMBOL to REFSEQ

\item \code{REFSEQ_to_SYMBOL}: convert from REFSEQ to SYMBOL
}}
\examples{
SYMBOL_to_ENTREZID(c("EGR1", "FOS"))
ENTREZID_to_SYMBOL(c("1958", "2353"))
SYMBOL_to_ENSEMBL(c("EGR1", "FOS"))
ENSEMBL_to_SYMBOL(c("ENSG00000120738", "ENSG00000170345"))
}

